# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AI Food Concierge API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * @nullable
 */
export type UserPreferencesSpiceLevel = typeof UserPreferencesSpiceLevel[keyof typeof UserPreferencesSpiceLevel] | null;


export const UserPreferencesSpiceLevel = {
  mild: 'mild',
  medium: 'medium',
  spicy: 'spicy',
  null: 'null',
} as const;

/**
 * @nullable
 */
export type UserPreferencesDiningPreference = typeof UserPreferencesDiningPreference[keyof typeof UserPreferencesDiningPreference] | null;


export const UserPreferencesDiningPreference = {
  'dine-in': 'dine-in',
  takeaway: 'takeaway',
  delivery: 'delivery',
  null: 'null',
} as const;

export interface UserPreferences {
  id: number;
  userId: string;
  /** @nullable */
  isHalal: boolean | null;
  /** @nullable */
  isVegetarian: boolean | null;
  /** @nullable */
  spiceLevel: UserPreferencesSpiceLevel;
  allergies: string[];
  favoriteCuisines: string[];
  /** @nullable */
  budgetRange: string | null;
  /** @nullable */
  diningPreference: UserPreferencesDiningPreference;
  frequentAreas: string[];
  onboardingCompleted: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface PreferencesInput {
  /** @nullable */
  isHalal?: boolean | null;
  /** @nullable */
  isVegetarian?: boolean | null;
  /** @nullable */
  spiceLevel?: string | null;
  allergies?: string[];
  favoriteCuisines?: string[];
  /** @nullable */
  budgetRange?: string | null;
  /** @nullable */
  diningPreference?: string | null;
  frequentAreas?: string[];
  onboardingCompleted?: boolean;
}

export type RestaurantBudgetRange = typeof RestaurantBudgetRange[keyof typeof RestaurantBudgetRange];


export const RestaurantBudgetRange = {
  'RM5-RM15': 'RM5-RM15',
  'RM15-RM30': 'RM15-RM30',
  'RM30+': 'RM30+',
} as const;

export interface Restaurant {
  id: number;
  name: string;
  cuisine: string;
  description: string;
  address: string;
  area: string;
  /** @nullable */
  latitude: number | null;
  /** @nullable */
  longitude: number | null;
  budgetRange: RestaurantBudgetRange;
  atmosphere: string[];
  diningOptions: string[];
  openingHours: string;
  isOpenNow: boolean;
  rating: number;
  reviewCount: number;
  photos: string[];
  popularDishes: string[];
  isHalal: boolean;
  isVegetarianFriendly: boolean;
  tags: string[];
  isFeatured: boolean;
  createdAt: string;
}

export interface RecommendationRequest {
  /** Natural language mood/craving description */
  mood: string;
  /** @nullable */
  budget?: string | null;
  /** @nullable */
  distance?: number | null;
  /** @nullable */
  cuisine?: string | null;
  /** @nullable */
  atmosphere?: string | null;
  /** @nullable */
  diningPreference?: string | null;
  /** IDs of restaurants to exclude (for shuffle) */
  excludeRestaurantIds?: number[];
}

export interface RecommendationResult {
  restaurant: Restaurant;
  /** Why this restaurant was recommended */
  matchReason: string;
  /** How the AI interpreted the mood */
  moodInterpretation: string;
  score: number;
  alternativeIds: number[];
}

export interface RecommendationHistory {
  id: number;
  restaurant: Restaurant;
  mood: string;
  matchReason: string;
  createdAt: string;
}

export interface FavoriteRestaurant {
  id: number;
  restaurant: Restaurant;
  createdAt: string;
}

export interface FavoriteInput {
  restaurantId: number;
}

export type DashboardDataTopCuisinesItem = {
  cuisine: string;
  count: number;
};

export interface DashboardData {
  totalFavorites: number;
  totalRecommendations: number;
  topCuisines: DashboardDataTopCuisinesItem[];
  recentRecommendations: RecommendationHistory[];
  favoriteRestaurants: FavoriteRestaurant[];
}

export type ListRestaurantsParams = {
cuisine?: string;
budget?: string;
atmosphere?: string;
distance?: number;
openNow?: boolean;
limit?: number;
};

export type GetRecommendationHistoryParams = {
limit?: number;
};