# lib/api-zod/src/generated/types/dashboardData.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AI Food Concierge API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardDataTopCuisinesItem } from './dashboardDataTopCuisinesItem';
import type { FavoriteRestaurant } from './favoriteRestaurant';
import type { RecommendationHistory } from './recommendationHistory';

export interface DashboardData {
  totalFavorites: number;
  totalRecommendations: number;
  topCuisines: DashboardDataTopCuisinesItem[];
  recentRecommendations: RecommendationHistory[];
  favoriteRestaurants: FavoriteRestaurant[];
}