# lib/api-zod/src/generated/types/preferencesInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AI Food Concierge API
 * OpenAPI spec version: 0.1.0
 */

export interface PreferencesInput {
  /** @nullable */
  isHalal?: boolean | null;
  /** @nullable */
  isVegetarian?: boolean | null;
  /** @nullable */
  spiceLevel?: string | null;
  allergies?: string[];
  favoriteCuisines?: string[];
  /** @nullable */
  budgetRange?: string | null;
  /** @nullable */
  diningPreference?: string | null;
  frequentAreas?: string[];
  onboardingCompleted?: boolean;
}