# lib/api-zod/src/generated/types/recommendationHistory.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AI Food Concierge API
 * OpenAPI spec version: 0.1.0
 */
import type { Restaurant } from './restaurant';

export interface RecommendationHistory {
  id: number;
  restaurant: Restaurant;
  mood: string;
  matchReason: string;
  createdAt: string;
}