# lib/api-zod/src/generated/types/recommendationRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AI Food Concierge API
 * OpenAPI spec version: 0.1.0
 */

export interface RecommendationRequest {
  /** Natural language mood/craving description */
  mood: string;
  /** @nullable */
  budget?: string | null;
  /** @nullable */
  distance?: number | null;
  /** @nullable */
  cuisine?: string | null;
  /** @nullable */
  atmosphere?: string | null;
  /** @nullable */
  diningPreference?: string | null;
  /** IDs of restaurants to exclude (for shuffle) */
  excludeRestaurantIds?: number[];
}