# lib/api-zod/src/generated/types/restaurant.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AI Food Concierge API
 * OpenAPI spec version: 0.1.0
 */
import type { RestaurantBudgetRange } from './restaurantBudgetRange';

export interface Restaurant {
  id: number;
  name: string;
  cuisine: string;
  description: string;
  address: string;
  area: string;
  /** @nullable */
  latitude: number | null;
  /** @nullable */
  longitude: number | null;
  budgetRange: RestaurantBudgetRange;
  atmosphere: string[];
  diningOptions: string[];
  openingHours: string;
  isOpenNow: boolean;
  rating: number;
  reviewCount: number;
  photos: string[];
  popularDishes: string[];
  isHalal: boolean;
  isVegetarianFriendly: boolean;
  tags: string[];
  isFeatured: boolean;
  createdAt: string;
}